refactor packages with updated flow-chart
