+++ src/main/java/com/nfs/academy/lrn/pattern/command/services/ICommand.java
@@ -1,6 +1,9 @@
-package com.nfs.academy.lrn.pattern.command;
+package com.nfs.academy.lrn.pattern.command.services;
 
 import java.io.Serializable;
+
+import com.nfs.academy.lrn.pattern.command.util.Command;
+import com.nfs.academy.lrn.pattern.command.entity.Device;
 
 public interface ICommand extends Serializable {
