+++ src/main/java/com/nfs/academy/lrn/pattern/factory/CommandFactory.java
@@ -1,11 +1,14 @@
 package com.nfs.academy.lrn.pattern.factory;
 
-import com.nfs.academy.lrn.pattern.command.ChannelImpl;
-import com.nfs.academy.lrn.pattern.command.Command;
-import com.nfs.academy.lrn.pattern.command.ICommand;
-import com.nfs.academy.lrn.pattern.command.PlayListImpl;
-import com.nfs.academy.lrn.pattern.command.PowerImpl;
-import com.nfs.academy.lrn.pattern.command.VolumeImpl;
+import java.util.HashMap;
+import java.util.Map;
+
+import com.nfs.academy.lrn.pattern.command.services.ChannelImpl;
+import com.nfs.academy.lrn.pattern.command.util.Command;
+import com.nfs.academy.lrn.pattern.command.services.ICommand;
+import com.nfs.academy.lrn.pattern.command.services.PlayListImpl;
+import com.nfs.academy.lrn.pattern.command.services.PowerImpl;
+import com.nfs.academy.lrn.pattern.command.services.VolumeImpl;
 
 public class CommandFactory {
 
@@ -13,18 +16,46 @@ public class CommandFactory {
         switch (type) {
             case ON:
             case OFF:
-                return new PowerImpl();
+//                return new PowerImpl();
+                return Cache.get(Cache.Key.POWER);
             case INCREASE_VOLUME:
             case DECREASE_VOLUME:
-                return new VolumeImpl();
+//                return new VolumeImpl();
+                return Cache.get(Cache.Key.VOLUME);
             case INCREMENT_CHANNEL:
             case DECREMENT_CHANNEL:
-                return new ChannelImpl();
+//                return new ChannelImpl();
+                return Cache.get(Cache.Key.CHANNEL);
             case ADD_FAVOURITE:
             case REMOVE_FAVOURITE:
-                return new PlayListImpl();
+//                return new PlayListImpl();
+                return Cache.get(Cache.Key.PLAYLIST);
             default:
                 throw new IllegalArgumentException("Could generate object for the given type: " + type);
+        }
+    }
+
+    // caching with flyweight
+    private static class Cache {
+        static final Map<Key, ICommand> lookup;
+
+        static {
+            lookup = new HashMap<>();
+            lookup.put(Key.POWER, new PowerImpl());
+            lookup.put(Key.VOLUME, new VolumeImpl());
+            lookup.put(Key.CHANNEL, new ChannelImpl());
+            lookup.put(Key.PLAYLIST, new PlayListImpl());
+        }
+
+        private enum Key {
+            POWER,
+            VOLUME,
+            CHANNEL,
+            PLAYLIST
+        }
+
+        static ICommand get(Key key) {
+            return lookup.get(key);
         }
     }
